fix(랭킹페이지): myRank int -> Long

+++ src/main/java/com/hh99/nearby/entity/Member.java
@@ -67,5 +67,5 @@ public class Member extends Timestamped{
     private List<MemberChallenge> memberChallengeList2;
 
     @Column
-    private int myRank;
+    private Long myRank;
 }
